Added new table to enable a default sort order for output
[Mod][Add]: Added to the Contract, Database Helper, and Content Provider
a second table & column: def_display_order. This column will allows the
pre-specified order defined in the xml to be honour at query time. 

[Mod]: Updated tests to honour the new column & foreign key constraint.
In addition, one enhanced the remaning tests to use a non-application
context (via ProviderTestCase2 & RenamingDelegatingContext). This allows
tests to be done without cross contaminating the database or content
provider.

[Mod]: Changed the ExchangeRateUpdateLoader to ensure that the sort
order is added on first run.
[Mod]: Changed selection projection for the CursorAdapter to exclude the
source currency & specify which _ID to use.

This also lays the groundwork for issue #8 (additional columns may be
used for custom sorts).